pyproject.toml
[project]
name = "melo-ai-generated-summaries"
version = "0.1.1"
description = "AI-powered patient summary generation service"
readme = "README.md"
requires-python = ">=3.13.2"
dependencies = [
    "asyncpg>=0.30.0",
    "cryptography>=44.0.2",
    "fastapi[standard]>=0.115.12",
    "openai>=1.68.2",
    "psycopg2-binary>=2.9.10",
    "pyjwt>=2.10.1",
    "requests>=2.32.3",
    "sqlacodegen>=3.0.0",
    "sqlalchemy>=2.0.39",
]
